pyproject.toml
[project]
name = "ai-economy-timelines"
version = "0.1.0"
description = "Phase 1+2: historical compute baseline and forward compute-capacity model"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "statsmodels>=0.14.6",
]

[project.scripts]
historical = "pipelines.historical:main"
supply = "pipelines.supply:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["model", "pipelines"]

[tool.pytest.ini_options]
testpaths = ["tests"]
